requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.127.0
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ijson==3.4.0.post0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.1.0
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.12
langsmith==0.4.53
numpy==2.4.0
openai==2.8.1
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
pymilvus==2.6.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.40.0
wheel==0.45.1
xxhash==3.6.0
zstandard==0.25.0
